NodeMove_2 tests
