+++ src/Tests/SenseNet.IntegrationTests/MsSqlTests/MsSqlDataProviderTests.cs
@@ -203,119 +203,5 @@ namespace SenseNet.IntegrationTests.MsSqlTests
 
         [TestMethod]
         public Task IntT_MsSql_DP_Schema_ExclusiveUpdate() { return TestCase.DP_Schema_ExclusiveUpdate(); }
-
-        /* ================================================================================================== Move */
-
-        [TestMethod]
-        [ExpectedException(typeof(InvalidOperationException))]
-        public void IntT_MsSql_DP_Move_SourceIsNotExist() { TestCase.DP_Move_SourceIsNotExist(); }
-        [TestMethod]
-        [ExpectedException(typeof(InvalidOperationException))]
-        public void IntT_MsSql_DP_Move_TargetIsNotExist() { TestCase.DP_Move_TargetIsNotExist(); }
-        [TestMethod]
-        [ExpectedException(typeof(ArgumentNullException))]
-        public void IntT_MsSql_DP_Move_MoveTo_Null() { TestCase.DP_Move_MoveTo_Null(); }
-        [TestMethod]
-        [ExpectedException(typeof(ArgumentNullException))]
-        public void IntT_MsSql_DP_Move_NullSourcePath() { TestCase.DP_Move_NullSourcePath(); }
-        [TestMethod]
-        [ExpectedException(typeof(InvalidOperationException))]
-        public void IntT_MsSql_DP_Move_InvalidSourcePath() { TestCase.DP_Move_InvalidSourcePath(); }
-        [TestMethod]
-        [ExpectedException(typeof(ArgumentNullException))]
-        public void IntT_MsSql_DP_Move_NullTargetPath() { TestCase.DP_Move_NullTargetPath(); }
-        [TestMethod]
-        [ExpectedException(typeof(InvalidOperationException))]
-        public void IntT_MsSql_DP_Move_InvalidTargetPath() { TestCase.DP_Move_InvalidTargetPath(); }
-        [TestMethod]
-        [ExpectedException(typeof(InvalidOperationException))]
-        public void IntT_MsSql_DP_Move_ToItsParent() { TestCase.DP_Move_ToItsParent(); }
-        [TestMethod]
-        [ExpectedException(typeof(InvalidOperationException))]
-        public void IntT_MsSql_DP_Move_ToItself() { TestCase.DP_Move_ToItself(); }
-        [TestMethod]
-        [ExpectedException(typeof(InvalidOperationException))]
-        public void IntT_MsSql_DP_Move_ToUnderItself() { TestCase.DP_Move_ToUnderItself(); }
-        [TestMethod]
-        [ExpectedException(typeof(NodeAlreadyExistsException))]
-        public void IntT_MsSql_DP_Move_TargetHasSameName() { TestCase.DP_Move_TargetHasSameName(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_NodeTreeToNode() { TestCase.DP_Move_NodeTreeToNode(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_SourceIsLockedByAnother() { TestCase.DP_Move_SourceIsLockedByAnother(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_SourceIsLockedByCurrent() { TestCase.DP_Move_SourceIsLockedByCurrent(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_LockedTarget_SameUser() { TestCase.DP_Move_LockedTarget_SameUser(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_PathBeforeAfter() { TestCase.DP_Move_PathBeforeAfter(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_MinimalPermissions() { TestCase.DP_Move_MinimalPermissions(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_SourceWithoutDeletePermission() { TestCase.DP_Move_SourceWithoutDeletePermission(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_TargetWithoutAddNewPermission() { TestCase.DP_Move_TargetWithoutAddNewPermission(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_MoreVersion() { TestCase.DP_Move_MoreVersion(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_WithAspect() { TestCase.DP_Move_WithAspect(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_LeafNodeToContentList() { TestCase.DP_Move_ContentList_LeafNodeToContentList(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_LeafNodeToContentListItem() { TestCase.DP_Move_ContentList_LeafNodeToContentListItem(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_NodeTreeToContentList() { TestCase.DP_Move_ContentList_NodeTreeToContentList(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_NodeTreeToContentListItem() { TestCase.DP_Move_ContentList_NodeTreeToContentListItem(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_NodeWithContentListToNode() { TestCase.DP_Move_ContentList_NodeWithContentListToNode(); }
-        [TestMethod]
-        [ExpectedException(typeof(InvalidOperationException))]
-        public void IntT_MsSql_DP_Move_ContentList_NodeWithContentListToContentList() { TestCase.DP_Move_ContentList_NodeWithContentListToContentList(); }
-        [TestMethod]
-        [ExpectedException(typeof(InvalidOperationException))]
-        public void IntT_MsSql_DP_Move_ContentList_NodeWithContentListToContentListItem() { TestCase.DP_Move_ContentList_NodeWithContentListToContentListItem(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListToNode() { TestCase.DP_Move_ContentList_ContentListToNode(); }
-        [TestMethod]
-        [ExpectedException(typeof(InvalidOperationException))]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListToContentList() { TestCase.DP_Move_ContentList_ContentListToContentList(); }
-        [TestMethod]
-        [ExpectedException(typeof(InvalidOperationException))]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListToContentListItem() { TestCase.DP_Move_ContentList_ContentListToContentListItem(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListTreeToNode() { TestCase.DP_Move_ContentList_ContentListTreeToNode(); }
-        [TestMethod]
-        [ExpectedException(typeof(InvalidOperationException))]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListTreeToContentList() { TestCase.DP_Move_ContentList_ContentListTreeToContentList(); }
-        [TestMethod]
-        [ExpectedException(typeof(InvalidOperationException))]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListTreeToContentListItem() { TestCase.DP_Move_ContentList_ContentListTreeToContentListItem(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListItemToNode() { TestCase.DP_Move_ContentList_ContentListItemToNode(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListItemToContentList() { TestCase.DP_Move_ContentList_ContentListItemToContentList(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListItemToContentListItem() { TestCase.DP_Move_ContentList_ContentListItemToContentListItem(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListItemTreeToNode() { TestCase.DP_Move_ContentList_ContentListItemTreeToNode(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListItemTreeToContentList() { TestCase.DP_Move_ContentList_ContentListItemTreeToContentList(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListItemTreeToContentListItem() { TestCase.DP_Move_ContentList_ContentListItemTreeToContentListItem(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListItemTree2ToNode() { TestCase.DP_Move_ContentList_ContentListItemTree2ToNode(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListItemTree2ToContentList() { TestCase.DP_Move_ContentList_ContentListItemTree2ToContentList(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListItemTree2ToContentListItem() { TestCase.DP_Move_ContentList_ContentListItemTree2ToContentListItem(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListItemToSameContentList() { TestCase.DP_Move_ContentList_ContentListItemToSameContentList(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListItemToSameContentListItem() { TestCase.DP_Move_ContentList_ContentListItemToSameContentListItem(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListItemTreeToSameContentList() { TestCase.DP_Move_ContentList_ContentListItemTreeToSameContentList(); }
-        [TestMethod]
-        public void IntT_MsSql_DP_Move_ContentList_ContentListItemTreeToSameContentListItem() { TestCase.DP_Move_ContentList_ContentListItemTreeToSameContentListItem(); }
     }
 }
